Change server system

+++ VoiceCraft.Core/Network/Packets/AudioPacket.cs
@@ -6,12 +6,12 @@ namespace VoiceCraft.Core.Network.Packets
     public class AudioPacket : VoiceCraftPacket
     {
         public override PacketType PacketType => PacketType.Audio;
-        public int Id { get; private set; }
+        public byte Id { get; private set; }
         public uint Timestamp { get; private set; }
         public int Length { get; private set; }
         public byte[] Data { get; private set; }
         
-        public AudioPacket(int id = 0, uint timestamp = 0, int length = 0, byte[]? data = null)
+        public AudioPacket(byte id = 0, uint timestamp = 0, int length = 0, byte[]? data = null)
         {
             Id = id;
             Timestamp = timestamp;
@@ -28,7 +28,7 @@ namespace VoiceCraft.Core.Network.Packets
 
         public override void Deserialize(NetDataReader reader)
         {
-            Id = reader.GetInt();
+            Id = reader.GetByte();
             Timestamp = reader.GetUInt();
             Length = reader.AvailableBytes;
             //Fuck no. we aren't allocating anything higher than the expected amount of bytes (WHICH SHOULD BE COMPRESSED!).
